Adjusted the HTML so that the canvases will be under each other if there isn't enough width (phone). Also fixed some of the logic.
Images smaller than 250x250 would appear distorted so I thought I'd fix the scale. As I was looking at the code, I realized that, although we're passing the right number of bytes to be adjusted, only the first 250 pixels had valid data because we copy the image data from the canvas which is displayed at 250x250. The rest of the bytes were 0. The logic has been adjusted to use a temporary canvas to get the original pixel data and pass that to the functions for modification. The renderModifiedImage function had to be adjusted in order to display the modified image at the proper scale.